backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-CORS
psycopg2-binary
